requirements.txt
selenium
pytest
pyyaml
requests
psycopg2-binary

